requirement.txt
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Flask-MySQL==1.5.1
Flask-Caching==1.9.0
PyMySQL==1.0.2
mysqlclient==2.0.1
redis==4.0.2
requests==2.24.0
retrying==1.3.3
jsonschema==3.2.0
PyJWT==1.4.2
pycryptodome==3.9.9
simplejson==3.17.2
jinja2==3.0.3
werkzeug==2.0.3
itsdangerous==1.1.0 # 2.0以后版本使用的为json，之前会尝试使用simplejson
SQLAlchemy==1.3.19
